Commit fixes

+++ pluq-meter-report/src/main/java/com/meterreport/pluqmeterreport/services/MeterValueService.java
@@ -1,5 +1,9 @@
 package com.meterreport.pluqmeterreport.services;
 
+
+import com.meterreport.pluqmeterreport.errors.customErrors.location.LocationNotFoundException;
+import com.meterreport.pluqmeterreport.errors.customErrors.meterValue.MeterValueAlreadyExistsException;
+import com.meterreport.pluqmeterreport.errors.customErrors.meterValue.MeterValueNotFoundException;
 import com.meterreport.pluqmeterreport.models.MeterValue;
 import com.meterreport.pluqmeterreport.repos.MeterValueRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,26 +24,53 @@ public class MeterValueService {
 
 
     public Optional<MeterValue> getMeterValueById(String meterValueId) {
-        return meterValueRepository.findById(meterValueId);
+        Optional<MeterValue> meterValue = meterValueRepository.findById(meterValueId);
+
+        if (meterValue.isEmpty()) {
+            throw new MeterValueNotFoundException("MeterValue not found by meterValueId: " + meterValueId);
+        }
+
+        return meterValue;
     }
 
     public List<MeterValue> getAllMeterValues() {
-        return meterValueRepository.findAll();
+        List<MeterValue> meterValueList = meterValueRepository.findAll();
+
+        if (meterValueList.isEmpty()) {
+            throw new MeterValueNotFoundException("Could not find any saved meterValues.");
+        }
+        return meterValueList;
     }
 
     public MeterValue saveMeterValue(MeterValue meterValue) {
+        if (meterValueRepository.findById(meterValue.getTransactionId()).isPresent()) {
+            throw new MeterValueAlreadyExistsException("MeterValue already exists with id: " + meterValue.getTransactionId());
+        }
+
         return meterValueRepository.save(meterValue);
     }
 
     public List<MeterValue> saveMeterValueList(List<MeterValue> meterValueList) {
+        for (MeterValue meterValue : meterValueList) {
+            if (meterValueRepository.findById(meterValue.getTransactionId()).isPresent()) {
+                throw new MeterValueAlreadyExistsException("MeterValue already exists with id: " + meterValue.getTransactionId());
+            }
+        }
         return meterValueRepository.saveAll(meterValueList);
     }
 
     public void deleteMeterValue(String meterValueId) {
+        if (meterValueRepository.findById(meterValueId).isEmpty()) {
+            throw new MeterValueNotFoundException("MeterValue not found by meterValueId: " + meterValueId);
+        }
+
         meterValueRepository.deleteById(meterValueId);
     }
 
     public void deleteAllMeterValues() {
+        if (meterValueRepository.findAll().isEmpty()) {
+            throw new LocationNotFoundException("Could not find any saved meterValues.");
+        }
         meterValueRepository.deleteAll();
     }
